+++ extensions/snap-jot/CHANGELOG.md
@@ -1,5 +1,9 @@
 # SnapJot Changelog
 
+## [New Commands] - 2024-07-13
+- Add "Search Text" command
+    - You can now search for text within your notes.
+    
 ## [Bug Fixes] - 2024-05-20
 - Line breaks would become incorrect when used in combination with other note-taking apps such as Thino by Obsidian.